Standalone go-importable RWTxt server.
Reworked project into a standalone go-importable RWTxt server.  This
loosens the coupling between web-server (/bin/rwtxt) logic and rwtxt
core functionality.

Other packages can now import and use rwtxt as a library, which opens
the possibility of incorporating and extending rwtxt inside larger
projects.

- Refactored singleton vars into new `RWTxt' type, along with affected
  functions.

- Renamed "src" package to commonly used convention "pkg".

- Migrated main.go to "rwtxt" command under cmd/rwtxt.

- Changed panics during RWTxt struct initialization to instead return an
  error.

- Added DB files and intermediate build artifacts to .gitignore.

- Removed trailing whitespace from Makefile and README.
