+++ Makefile
@@ -6,7 +6,7 @@ build:
 	go install -v github.com/tdewolff/minify/cmd/minify
 	go get -d -v github.com/jteeuwen/go-bindata/...
 	go install -v github.com/jteeuwen/go-bindata/go-bindata
-	rm -rf assets 
+	rm -rf assets
 	cp -r static assets
 	cd assets && gzip -9 -r *
 	cp templates/main.html assets/main.html
@@ -26,11 +26,11 @@ build:
 	# gzip -9 -c static/img/logo.png  > assets/logo.png
 	# cp -r static/img/favicon assets/
 	# cd assets/favicon && gzip -9 *
-	go-bindata -nocompress assets assets/img assets/js assets/css assets/img/favicon
-	go build -v --tags "fts4" ${LDFLAGS}
+	go-bindata -pkg rwtxt -nocompress assets assets/img assets/js assets/css assets/img/favicon
+	go get -v --tags "fts4" ${LDFLAGS} ./...
 
 run: build
-	./rwtxt
+	$(GOPATH)/bin/rwtxt
 
 dev:
 	rerun make run
@@ -39,5 +39,5 @@ release:
 	docker pull karalabe/xgo-latest
 	go get github.com/karalabe/xgo
 	mkdir -p bin
-	xgo -go "1.10.1" -dest bin ${LDFLAGS} -targets linux/amd64,linux/arm-6,darwin/amd64,windows/amd64 github.com/schollz/rwtxt
+	xgo -go $(shell go version) -dest bin ${LDFLAGS} -targets linux/amd64,linux/arm-6,darwin/amd64,windows/amd64 github.com/schollz/rwtxt/cmd/rwtxt
 	# cd bin && upx --brute kiki-linux-amd64
